remove entrypoint volumemapping
The entrypoint is different in initial images (nginx < 1.24).
The image should have the correct entrypoint. Both entrypoints are in the epadjs folder for reference and incase it is needed.
The mapping can be done with adding following line to volumes section in epadjs module in docker-compose.yml (Replace entrypoint.sh with the appropriate name according to the nginx version)
- ./epadjs/entrypoint.sh:/usr/local/bin/entrypoint